cambios en fragment maquina y añadir activity maquina

+++ app/src/main/java/com/multimedia/aes/gestnet_nucleo/nucleo/InfoArticulos.java
@@ -86,7 +86,7 @@ public class InfoArticulos  extends AppCompatActivity implements View.OnClickLis
       //ivFoto.setImageResource(articulo.getImagen());
         tvTitulo.setText(articulo.getNombre_articulo());
 
-        tvStock.setText(String.valueOf(articulo.getStock()));
+        tvStock.setText(String.valueOf((int)articulo.getStock()));
         if(articulo.getStock()<1){
             btnAñadirMaterial.setClickable(false);
             btnAñadirMaterial.setEnabled(false);
